install readme.md

+++ install.sh
@@ -1,15 +1,26 @@
 #!/usr/bin/env bash
 
-# install dein.vim
-curl -fsSL https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh | bash -s ./dein
+exists_cmd() {
+  command -v $1 >/dev/null 2>&1
+}
 
 # install pip3
-if hash apt-get 2>/dev/null; then
-  sudo apt-get install python3-pip
-elif hash brew 2>/dev/null; then
-  brew install python3
+if exists_cmd apt-get; then
+  if ! exists_cmd pip3; then
+    sudo apt-get install python3-pip
+  fi
+elif exists_cmd brew; then
+  if ! exists_cmd pip3; then
+    brew install python3
+  fi
 fi
 
-# install python yaml and neovim
-pip3 install pyyaml
-pip3 install neovim
+if exists_cmd pip3; then
+  # install python yaml and neovim
+  pip3 install pyyaml
+  pip3 install neovim
+else
+  echo 'manual install pyyaml and neovim please'
+fi
+
+./install-dein.sh
